requirements.txt
flask==2.0.1
pymongo==4.3.3
python-dotenv==0.19.0
twilio==7.7.0
requests==2.26.0
numpy==1.21.2
tensorflow==2.6.0
dnspython==2.2.1
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1 